account/pyproject.toml
[tool.poetry]
name = "account"
version = "0.1.0"
description = ""
authors = ["Sweepyd1 <dima10yashin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
fastapi = {extras = ["all"], version = "^0.115.0"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
uvicorn = "^0.31.0"

alembic = "^1.13.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
aio-pika = "^9.4.3"
aiohttp = "^3.10.9"
psycopg2-binary = "2.9.6"
nest-asyncio = "^1.6.0"
gunicorn = "^23.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
